Problem: baker-stats: Leaving outdated blocks around
We always store and parse the head of the chain, but unless it happens
to be the last block of a cycle, it becomes irrelevant when the next
block comes around.

'blocks' holds the relevant blocks -- the current head and the last
blocks of all preceding cycles.

Solution: At the end of baker-stats, remove all cached blocks not in 'blocks'.