Problem: Rewards are shown for the cycle they were earned
As a staker, I want to know when I will get my money. When it was
earned is a technical detail.

Solution: Record rewards for the cycle the will be paid out.

Which cycle directory the file is in doesn't matter, what matters is
the 'cycle' field. All rewards.json files are just concatenated
anyway.