Problem: Baker account is compiled into the stats script.
You cannot trivially run the stats on an arbitrary account.

Solution: Split out the stats script from the stats service.

Still not super convenient, but at least you can read the service,
read the service script, and then find the stats script and run it
with your choice of output dir and account address.

It's a first step.