Problem: accuser baker endorser: old protocol
Solution: Run the binaries for the new protocol.

The old protocol is safe to run in parallel, the services will become
inactive at fork time. We'll remove them again when this has happened.

I'm adding some copy-paste for this emergency fix. Next time this
happens, we'll have something neater in store.